optimization for saving logs
